Move all all the code from App.js to the new componen Home.js, and import the NavigationContainer component in new app.js
